+++ src/Locations/ui/TravelAgencyLocation.tsx
@@ -32,21 +32,16 @@ export class TravelAgencyLocation extends React.Component<IProps, any> {
     }
 
     render() {
-        const travelBtns: React.ReactNode[] = [];
-        for (const key in CityName) {
-            const city = CityName[key];
+        const thisTravelAgencyLocation = this;
 
-            // Skip current city
-            if (city === this.props.p.city) { continue; }
-
-            travelBtns.push(
-                <StdButton
-                    key={city}
-                    onClick={createTravelPopup.bind(null, city, this.props.travel)}
-                    style={this.btnStyle}
-                    text={`Travel to ${city}`}
-                />
-            )
+        function LocationLetter(props: any) {
+            if(props.city !== thisTravelAgencyLocation.props.p.city) {
+                return <span className='tooltip' style={{color: 'blue', whiteSpace: 'nowrap', margin: '0px', padding: '0px'}} onClick={createTravelPopup.bind(null, props.city, thisTravelAgencyLocation.props.travel)}>
+                    <span className='tooltiptext'>{props.city}</span>
+                    {props.city[0]}
+                </span>
+            }
+            return <span>{props.city[0]}</span>
         }
 
         return (
@@ -55,7 +50,28 @@ export class TravelAgencyLocation extends React.Component<IProps, any> {
                     From here, you can travel to any other city! A ticket
                     costs {numeralWrapper.formatMoney(CONSTANTS.TravelCost)}
                 </p>
-                {travelBtns}
+<pre>               ,_   .  ._. _.  .</pre>
+<pre>           , _-\','|~\~      ~/      ;-'_   _-'     ,;_;_,    ~~-</pre>
+<pre>  /~~-\_/-'~'--' \~~| ',    ,'      /  / ~|-_\_/~/~      ~~--~~~~'--_</pre>
+<pre>  /              ,/'-/~ '\ ,' _  , '<LocationLetter city='Volhaven' />,'|~                   ._/-, /~</pre>
+<pre>  ~/-'~\_,       '-,| '|. '   ~  ,\ /'~                /    /_  /~</pre>
+<pre>.-~      '|        '',\~|\       _\~     ,_  ,     <LocationLetter city='Chongqing' />         /,</pre>
+<pre>          '\     <LocationLetter city='Sector-12' />  /'~          |_/~\\,-,~  \ "         ,_,/ |</pre>
+<pre>           |       /            ._-~'\_ _~|              \ ) <LocationLetter city='New Tokyo' /></pre>
+<pre>            \   __-\           '/      ~ |\  \_          /  ~</pre>
+<pre>  .,         '\ |,  ~-_      - |          \\_' ~|  /\  \~ ,</pre>
+<pre>               ~-_'  _;       '\           '-,   \,' /\/  |</pre>
+<pre>                 '\_,~'\_       \_ _,       /'    '  |, /|'</pre>
+<pre>                   /     \_       ~ |      /         \  ~'; -,_.</pre>
+<pre>                   |       ~\        |    |  ,        '-_, ,; ~ ~\</pre>
+<pre>                    \,   <LocationLetter city='Aevum' />  /        \    / /|            ,-, ,   -,</pre>
+<pre>                     |    ,/          |  |' |/          ,-   ~ \   '.</pre>
+<pre>                    ,|   ,/           \ ,/              \   <LocationLetter city='Ishima' />   |</pre>
+<pre>                    /    |             ~                 -~~-, /   _</pre>
+<pre>                    | ,-'                                    ~    /</pre>
+<pre>                    / ,'                                      ~</pre>
+<pre>                    ',|  ~</pre>
+<pre>                      ~'</pre>
             </div>
         )
     }