+++ src/Netscript/RamCostGenerator.ts
@@ -202,7 +202,9 @@ export const RamCosts: IMap<any> = {
     getAugmentationsFromFaction: () => RamCostConstants.ScriptSingularityFn3RamCost,
     getAugmentationPrereq: () => RamCostConstants.ScriptSingularityFn3RamCost,
     getAugmentationCost: () => RamCostConstants.ScriptSingularityFn3RamCost,
+    getAugmentationStats: () => RamCostConstants.ScriptSingularityFn3RamCost,
     purchaseAugmentation: () => RamCostConstants.ScriptSingularityFn3RamCost,
+    softReset: () => RamCostConstants.ScriptSingularityFn3RamCost,
     installAugmentations: () => RamCostConstants.ScriptSingularityFn3RamCost,
 
     // Gang API