+++ src/ScriptEditor/AceNetscriptMode.js
@@ -91,7 +91,7 @@ let NetscriptFunctions =
     "createProgram|commitCrime|getCrimeChance|getOwnedAugmentations|"          +
     "getOwnedSourceFiles|getAugmentationsFromFaction|"                         +
     "getAugmentationPrereq|getAugmentationCost|purchaseAugmentation|"          +
-    "installAugmentations|"                                                    +
+    "softReset|installAugmentations|getAugmentationStats|"                      +
 
     // TIX API
     "getStockPrice|getStockPosition|getStockSymbols|getStockMaxShares|"        +
@@ -103,7 +103,7 @@ let NetscriptFunctions =
     // Hacknet Node API
     "hacknet|numNodes|purchaseNode|getPurchaseNodeCost|getNodeStats|"          +
     "upgradeLevel|upgradeRam|upgradeCore|upgradeCache|getLevelUpgradeCost|"    +
-    "getRamUpgradeCost|getCoreUpgradeCost|getCacheUpgradeCost|"                +
+    "getRamUpgradeCost|getCoreUpgradeCost|getCacheUpgradeCost|maxNumNodes|"    +
 
     // Gang API
     "gang|"                                                                    +